+++ packages/react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [6.82.1](https://github.com/factorialco/f0/compare/f0-react-v6.82.0...f0-react-v6.82.1) (2026-09-02)
+
+
+### Bug Fixes
+
+* **ci:** count .spec.ts(x) files as unit tests ([#5364](https://github.com/factorialco/f0/issues/5364)) ([75a3cc3](https://github.com/factorialco/f0/commit/75a3cc326b5c07de89eb96a9d337bddf067c334f))
+
+## [6.82.0](https://github.com/factorialco/f0/compare/f0-react-v6.81.3...f0-react-v6.82.0) (2026-09-02)
+
+
+### Features
+
+* **SlotWidget:** let a row's second line go critical ([#5360](https://github.com/factorialco/f0/issues/5360)) ([17f908f](https://github.com/factorialco/f0/commit/17f908fc91187fae0e6020fe74ee9358a7107cf7))
+
+## [6.81.3](https://github.com/factorialco/f0/compare/f0-react-v6.81.2...f0-react-v6.81.3) (2026-09-02)
+
+
+### Bug Fixes
+
+* **SlotWidget:** give View more the same gap below it as beside it ([#5363](https://github.com/factorialco/f0/issues/5363)) ([c35aa9b](https://github.com/factorialco/f0/commit/c35aa9b91d7c0dd2ff12eec190097dd94fee4652))
+
 ## [6.81.2](https://github.com/factorialco/f0/compare/f0-react-v6.81.1...f0-react-v6.81.2) (2026-09-02)